requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.2
django-cloudinary-storage==0.3.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
qrcode==8.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
